+++ packages/api/tests/lib/test-granule-remove-from-cmr.js
@@ -1,0 +1,196 @@
+'use strict';
+
+const test = require('ava');
+const sinon = require('sinon');
+const cryptoRandomString = require('crypto-random-string');
+
+const awsServices = require('@cumulus/aws-client/services');
+const StepFunctions = require('@cumulus/aws-client/StepFunctions');
+const launchpad = require('@cumulus/launchpad-auth');
+const { randomString } = require('@cumulus/common/test-utils');
+const { CMR } = require('@cumulus/cmr-client');
+const { DefaultProvider } = require('@cumulus/common/key-pair-provider');
+const {
+  generateLocalTestDb,
+  localStackConnectionEnv,
+} = require('@cumulus/db');
+
+const Granule = require('../../models/granules');
+const { fakeGranuleFactoryV2 } = require('../../lib/testUtils');
+const { unpublishGranule } = require('../../lib/granule-remove-from-cmr');
+const { migrationDir } = require('../../../../lambdas/db-migration');
+
+const testDbName = `granules_${cryptoRandomString({ length: 10 })}`;
+
+let fakeExecution;
+let sandbox;
+let testCumulusMessage;
+
+test.before(async (t) => {
+  process.env = {
+    ...process.env,
+    ...localStackConnectionEnv,
+    PG_DATABASE: testDbName,
+  };
+
+  process.env.GranulesTable = randomString();
+  await new Granule().createTable();
+
+  testCumulusMessage = {
+    cumulus_meta: {
+      execution_name: randomString(),
+      state_machine: 'arn:aws:states:us-east-1:123456789012:stateMachine:HelloStateMachine',
+      workflow_start_time: Date.now(),
+    },
+    meta: {
+      collection: {
+        name: randomString(),
+        version: randomString(),
+      },
+      provider: {
+        host: randomString(),
+        protocol: 's3',
+      },
+      status: 'completed',
+    },
+    payload: {
+      granules: [
+        {
+          granuleId: randomString(),
+          sync_granule_duration: 123,
+          post_to_cmr_duration: 456,
+          files: [],
+        },
+      ],
+    },
+  };
+
+  sandbox = sinon.createSandbox();
+
+  fakeExecution = {
+    input: JSON.stringify(testCumulusMessage),
+    startDate: new Date(Date.UTC(2019, 6, 28)),
+    stopDate: new Date(Date.UTC(2019, 6, 28, 1)),
+  };
+  sandbox.stub(StepFunctions, 'describeExecution').resolves(fakeExecution);
+
+  // Store the CMR password
+  process.env.cmr_password_secret_name = randomString();
+  await awsServices.secretsManager().createSecret({
+    Name: process.env.cmr_password_secret_name,
+    SecretString: randomString(),
+  }).promise();
+
+  // Store the launchpad passphrase
+  process.env.launchpad_passphrase_secret_name = randomString();
+  await awsServices.secretsManager().createSecret({
+    Name: process.env.launchpad_passphrase_secret_name,
+    SecretString: randomString(),
+  }).promise();
+
+  const { knex } = await generateLocalTestDb(testDbName, migrationDir);
+  t.context.knex = knex;
+});
+
+test.beforeEach((t) => {
+  t.context.granuleModel = new Granule();
+});
+
+test.after.always(async () => {
+  await awsServices.secretsManager().deleteSecret({
+    SecretId: process.env.cmr_password_secret_name,
+    ForceDeleteWithoutRecovery: true,
+  }).promise();
+  await awsServices.secretsManager().deleteSecret({
+    SecretId: process.env.launchpad_passphrase_secret_name,
+    ForceDeleteWithoutRecovery: true,
+  }).promise();
+  await new Granule().deleteTable();
+  sandbox.restore();
+});
+
+test('unpublishGranule() removing a granule from CMR fails if the granule is not in CMR', async (t) => {
+  const granule = fakeGranuleFactoryV2({ published: false });
+
+  await awsServices.dynamodbDocClient().put({
+    TableName: process.env.GranulesTable,
+    Item: granule,
+  }).promise();
+
+  try {
+    await unpublishGranule(t.context.knex, granule);
+  } catch (error) {
+    t.is(error.message, `Granule ${granule.granuleId} is not published to CMR, so cannot be removed from CMR`);
+  }
+});
+
+test.serial('removing a granule from CMR passes the granule UR to the cmr delete function', async (t) => {
+  sinon.stub(
+    DefaultProvider,
+    'decrypt'
+  ).callsFake(() => Promise.resolve('fakePassword'));
+
+  sinon.stub(
+    CMR.prototype,
+    'deleteGranule'
+  ).callsFake((granuleUr) => Promise.resolve(t.is(granuleUr, 'granule-ur')));
+
+  sinon.stub(
+    CMR.prototype,
+    'getGranuleMetadata'
+  ).callsFake(() => Promise.resolve({ title: 'granule-ur' }));
+
+  try {
+    const granule = fakeGranuleFactoryV2();
+
+    await awsServices.dynamodbDocClient().put({
+      TableName: process.env.GranulesTable,
+      Item: granule,
+    }).promise();
+
+    await unpublishGranule(t.context.knex, granule);
+  } finally {
+    CMR.prototype.deleteGranule.restore();
+    DefaultProvider.decrypt.restore();
+    CMR.prototype.getGranuleMetadata.restore();
+  }
+});
+
+test.serial('removing a granule from CMR succeeds with Launchpad authentication', async (t) => {
+  process.env.cmr_oauth_provider = 'launchpad';
+  const launchpadStub = sinon.stub(launchpad, 'getLaunchpadToken').callsFake(() => randomString());
+
+  sinon.stub(
+    DefaultProvider,
+    'decrypt'
+  ).callsFake(() => Promise.resolve('fakePassword'));
+
+  sinon.stub(
+    CMR.prototype,
+    'deleteGranule'
+  ).callsFake((granuleUr) => Promise.resolve(t.is(granuleUr, 'granule-ur')));
+
+  sinon.stub(
+    CMR.prototype,
+    'getGranuleMetadata'
+  ).callsFake(() => Promise.resolve({ title: 'granule-ur' }));
+
+  try {
+    const granule = fakeGranuleFactoryV2();
+
+    await awsServices.dynamodbDocClient().put({
+      TableName: process.env.GranulesTable,
+      Item: granule,
+    }).promise();
+
+    await unpublishGranule(t.context.knex, granule);
+
+    t.is(launchpadStub.calledOnce, true);
+  } finally {
+    process.env.cmr_oauth_provider = 'earthdata';
+    launchpadStub.restore();
+    CMR.prototype.deleteGranule.restore();
+    DefaultProvider.decrypt.restore();
+    CMR.prototype.getGranuleMetadata.restore();
+  }
+});